Separated common start functions into sosmurf.util library
