+++ python/sosmurf/SmurfStreamer/__init__.py
@@ -1,0 +1,2 @@
+
+from sosmurf._SmurfStreamer import SmurfStreamer
