+++ python/sosmurf/__init__.py
@@ -1,0 +1,2 @@
+import sosmurf.util
+from sosmurf.SmurfStreamer import SmurfStreamer
